+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 - 2019-10-22
+**Added**
+
+* Custom user picture meta key.
+* Override `get_avatar()` to work with WP's image sizes.
+
 ## 1.0.1 - 2017-10-09
 **Fixed**